Working on navbar, landing page and promotion
