+++ client/components/Navbar.jsx
@@ -1,27 +1,39 @@
 import React from 'react';
 import styled from 'styled-components';
+import { Link } from "react-router-dom";
 
 const Nav = styled.div`
   display: flex;
   text-decoration: none;
   flex-direction: row;
-  align-items: left;
-  justify-content: left;
-  background-color: transparent;
+  align-items: center;
+  justify-content: flex-start;
+  background-color: #192A34;
   height: 6vh;
   padding-right: 1px;
   padding-top: 1.5vh;
 `;
-  // padding-left: 10%;
-  // margin-left: -10%;
-  // margin-right: -8%;
-  // margin-top: -9px;
+
+const NavLink = styled(Link)`
+  font-family: Helvetica;
+  font-size: 14px;
+  padding-right: 1vw;
+  text-decoration: none;
+  color: white;
+  NavLink:visited {
+    color: white;
+  }
+  &:hover ${NavLink} {
+    color: #E9EB9E;
+    border: none;
+  }
+`;
 
 const Job = styled.div`
   font-family: Helvetica;
   color: #E9EB9E;
   font-size: 5vh;
-  padding-left: 10vw;
+  padding-left: 1vw;
 `;
 
 const Site = styled.div`
@@ -31,29 +43,15 @@ const Site = styled.div`
   padding-right: 55%;
 `;
 
-const NavButtons = styled.div`
-  font-family: Helvetica;
-  color: white;
-  font-size: 24px;
-  padding-right: 2vw;
-  padding-top: 2.5vh;
-  white-space: nowrap;
-  &:hover ${NavButtons} {
-    color: #E9EB9E;
-    pointer: cursor;
-}`
-
 const Navbar = () => {
   return (
-    <div>
-      <Nav>
-        <Job>Job</Job>
-        <Site>Site</Site>
-        <NavButtons>Find Jobs</NavButtons>
-        <NavButtons>Employers</NavButtons>
-        <NavButtons>Log In</NavButtons>
-      </Nav>
-    </div>
+    <Nav>
+      <Job>Job</Job>
+      <Site>Site</Site>
+      <NavLink to="/">Find Jobs</NavLink>
+      <NavLink to="/profile">Post Jobs</NavLink>
+      <NavLink to="/login">Log In</NavLink>
+    </Nav>
   )
 }
